+++ lib/angular/directives/fileDetailController.js
@@ -4,7 +4,7 @@
  * @param mediator - mediator service to publish file:close topic, client app can subscribe to this topic.
  */
 
-function FieldDetailController($scope, mediator){
+function fileDetailController($scope, mediator){
   var self = this;
   self.file = $scope.file;
   self.options = $scope.displayOptions || {};
@@ -16,4 +16,6 @@ function FieldDetailController($scope, mediator){
   };
 }
 
-angular.module('wfm.file.directives').controller('FieldDetailController', ['$scope','mediator', FieldDetailController]);
+angular.module('wfm.file.directives').controller('fileDetailController', ['$scope','mediator', fileDetailController]);
+
+module.exports = 'wfm.file.directives.fileDetailController';
