+++ lib/angular/directives/fileDetailDirective.js
@@ -1,5 +1,5 @@
 /**
- * Detail view of a file, passed in is file object and options.
+ * Detail view o a file, passed in is file object and options.
  * Options help to chose which details we want to render
  * options e.g. {id: true, name: true, uid: true, owner: true, preview:true};
  */
@@ -14,7 +14,7 @@ angular.module('wfm.file.directives').directive('fileDetail', function($template
         file: '=',
         displayOptions: '='
       },
-      controller: 'FieldDetailController',
+      controller: 'fileDetailController',
       controllerAs: 'ctrl'
     };
 });
