+++ lib/angular/file-ng.js
@@ -6,7 +6,7 @@ var client = require('../file'),
 module.exports = 'wfm.file.service';
 
 angular.module('wfm.file.service', [
-  require('./directives')
+  require('./fileDirectives')
 ])
 
 .factory('fileClient', function($q) {
