+++ lib/angular/fileDirectives.js
